:star: FT_67: Tender Type CSV Mapping for Catapult for EBC - Added new field mapping for `Tender Type`
- Created `.sql` file for field mapping
- Added new functions in migration trait
- Created new `DisplayState` enum
- Modified migration for `payment_method_settings`

+++ database/migrations/2023_02_02_160913_create_cdis_payment_tender_type_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\DB;
 
 class CreateCdisPaymentTenderTypeTable extends Migration
 {
@@ -13,10 +14,20 @@ class CreateCdisPaymentTenderTypeTable extends Migration
      */
     public function up()
     {
-        Schema::create('cdis_payment_tender_type', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->timestamps();
-        });
+        if (! Schema::hasTable('cdis_payment_tender_type')) {
+            Schema::create('cdis_payment_tender_type', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->unsignedBigInteger('bid')->index()->unique();
+                $table->string('name', 128);
+                $table->tinyInteger('status')->default(\App\Enums\Status::ACTIVE);
+                $table->tinyInteger('has_field_description')->default(\App\Enums\DisplayState::NO);
+                $table->unsignedBigInteger('created_by')->nullable();
+                $table->unsignedBigInteger('updated_by')->nullable();
+                $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
+                $table->timestamp('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
+                $table->softDeletes();
+            });
+        }
     }
 
     /**
@@ -26,6 +37,8 @@ class CreateCdisPaymentTenderTypeTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('cdis_payment_tender_type');
+        if (Schema::hasTable('cdis_payment_tender_type')) {
+            Schema::dropIfExists('cdis_payment_tender_type');
+        }
     }
 }
