+++ database/migrations/2023_02_02_160953_create_cdis_payment_charge_type_table.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\DB;
 
 class CreateCdisPaymentChargeTypeTable extends Migration
 {
@@ -13,10 +14,19 @@ class CreateCdisPaymentChargeTypeTable extends Migration
      */
     public function up()
     {
-        Schema::create('cdis_payment_charge_type', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->timestamps();
-        });
+        if (! Schema::hasTable('cdis_payment_charge_type')) {
+            Schema::create('cdis_payment_charge_type', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->unsignedBigInteger('bid')->index()->unique();
+                $table->string('name', 128);
+                $table->tinyInteger('status')->default(\App\Enums\Status::ACTIVE);
+                $table->unsignedBigInteger('created_by')->nullable();
+                $table->unsignedBigInteger('updated_by')->nullable();
+                $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
+                $table->timestamp('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
+                $table->softDeletes();
+            });
+        }
     }
 
     /**
@@ -26,6 +36,8 @@ class CreateCdisPaymentChargeTypeTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('cdis_payment_charge_type');
+        if (Schema::hasTable('cdis_payment_charge_type')) {
+            Schema::dropIfExists('cdis_payment_charge_type');
+        }
     }
 }
